Move remaining classes to registry project
