+++ core/src/main/java/de/unistuttgart/iaas/amyassist/amy/core/CoreServiceProviderLoader.java
@@ -25,8 +25,8 @@ package de.unistuttgart.iaas.amyassist.amy.core;
 
 import java.util.Properties;
 
-import de.unistuttgart.iaas.amyassist.amy.core.registry.contact.ContactRegistryImpl;
-import de.unistuttgart.iaas.amyassist.amy.core.registry.location.LocationRegistryImpl;
+import de.unistuttgart.iaas.amyassist.amy.registry.ContactRegistryImpl;
+import de.unistuttgart.iaas.amyassist.amy.registry.LocationRegistryImpl;
 import org.slf4j.Logger;
 
 import de.unistuttgart.iaas.amyassist.amy.core.configuration.ConfigurationLoaderImpl;
